sur/requirements.txt
numpy
pillow
imageio
imageio-ffmpeg
scipy
tqdm
flask
gunicorn
